Only create env files if they dont exist
This prevents the need to keep writing secrets in files
If file exists, it will not be overwritten
To change environments, delete env files and run script

+++ MigrantHub/run_local.sh
@@ -1,7 +1,12 @@
 #!/usr/bin/env bash
 
-cp ./server/.env_local ./server/.env
-cp ./client/.env_local ./client/.env.local
+if [ ! -f ./server/.env ]; then
+  cp ./server/.env_local ./server/.env
+fi
+
+if [ ! -f ./client/.env.local ]; then
+  cp ./client/.env_local ./client/.env.local
+fi
 
 cd ./server
 npm run dev
